+++ sim-administration/simmanager/src/main/java/org/ostelco/simcards/admin/SimAdministrationModule.kt
@@ -60,7 +60,7 @@ class SimAdministrationModule : PrimeModule {
 
 
         var hssAdapters =
-                HssAdapterCache(
+                HssAdapterManager(
                         httpClient = httpClient,
                         simInventoryDAO = this.DAO,
                         hssConfigs = config.hssVendors)
